Minor styling tweaks in header...

+++ src/components/header/Header.jsx
@@ -20,8 +20,12 @@ const Header = () => {
     }, [])
 
     const logoutHandler = () =>{
-        localStorage.clear();
-        navigateTo("/");
+
+        const confirmFlag = window.confirm("Are you sure to log out?")
+        if (confirmFlag){
+            localStorage.clear();
+            navigateTo("/");
+        }
     }
 
     return (
@@ -37,8 +41,8 @@ const Header = () => {
                         <div className="col grow flex justify-between">
                             <div className="menu-wrapper text-sm md:text-lg text-gray-300">
                                 <ul className='menus flex flex-col gap-2 md:flex-row md:gap-8'>
-                                    <li className='menu'><NavLink to={`/contacts`}>All Contacts</NavLink></li>
-                                    <li className='menu'><NavLink to={`/add-contact`}>Add Contact</NavLink></li>
+                                    <li className='menu hover:font-semibold hover:text-white'><NavLink to={`/contacts`}>All Contacts</NavLink></li>
+                                    <li className='menu hover:font-semibold hover:text-white'><NavLink to={`/add-contact`}>Add Contact</NavLink></li>
                                 </ul>
                             </div>
                             <div className="user-menu-wrapper text-sm md:text-lg text-gray-300">
@@ -46,18 +50,18 @@ const Header = () => {
                                     isLoggedin?
                                         (
                                             <ul className='menus flex flex-col gap-2 md:flex-row md:gap-8'>
-                                                <li className='menu'><NavLink to={`/`}>{loggedInUser}</NavLink></li>
+                                                <li className='menu hover:font-semibold hover:text-white'><NavLink to={`/`}>{loggedInUser}</NavLink></li>
                                                 {/* <li className='menu'><NavLink to={`/`}>Logout</NavLink></li> */}
                                                 <li className='menu'>
-                                                    <button onClick={logoutHandler}>Logout</button>
+                                                    <button onClick={logoutHandler} className='hover:font-semibold hover:text-white'>Logout</button>
                                                 </li>
                                             </ul>
                                         )
                                     :
                                         (
                                         <ul className='menus flex flex-col gap-2 md:flex-row md:gap-8'>
-                                            <li className='menu'><NavLink to={`/login`}>Login</NavLink></li>
-                                            <li className='menu'><NavLink to={`/register`}>Register</NavLink></li>
+                                            <li className='menu hover:font-semibold hover:text-white'><NavLink to={`/login`}>Login</NavLink></li>
+                                            <li className='menu hover:font-semibold hover:text-white'><NavLink to={`/register`}>Register</NavLink></li>
                                         </ul>
                                         )
                                     }
